+++ src/components/graphs/PerformanceChart.tsx
