+++ src/components/graphs/PerformanceDisplay.tsx
