+++ src/components/graphs/PeriodBar.tsx
