+++ src/pages/application.tsx
@@ -1,7 +1,7 @@
-import Title from '../components/appliance/Title';
-import { HorizontalLine } from '../components/appliance/HorizontalLine';
-import SubTitle from '../components/appliance/SubTitle';
-import { Paragraph } from '../components/appliance/Paragraph';
+import Title from '../components/all/Title';
+import { HorizontalLine } from '../components/all/HorizontalLine';
+import SubTitle from '../components/all/SubTitle';
+import { Paragraph } from '../components/all/Paragraph';
 import { applicationText } from '../content';
 
 const blah = [
@@ -10,12 +10,12 @@ const blah = [
   'Hvordan midlene går onlinere til gode',
   'Aktivitetsplan',
   'Budsjett',
-]
+];
 
 export default function SoknadPage() {
   return (
     <div className="flex flex-col items-center justify-center w-full h-full">
-      <div className='w-full p-12 sm:w-4/5'>
+      <div className="w-full p-12 sm:w-4/5">
         <Title title="Søke om støtte" />
         <HorizontalLine />
         <SubTitle title="Hvordan søke om støtte?" />
@@ -23,11 +23,11 @@ export default function SoknadPage() {
         <Paragraph text="Søknader sendes til fondet@online.ntnu.no" />
         <Paragraph text="Dette skal være med i søknaden:" />
         <ul className="mt-8">
-          {
-            blah.map((item) => (
-              <li className="mx-8 my-4 text-xl list-disc sm:mx-12 md:mx-16">{item}</li>
-            ))
-          }
+          {blah.map((item) => (
+            <li className="mx-8 my-4 text-xl list-disc sm:mx-12 md:mx-16">
+              {item}
+            </li>
+          ))}
         </ul>
       </div>
     </div>
